Create frequency_divider
In this example, the frequency divider uses a register counter that increments on every positive edge of the input clock clk_in. When the counter reaches the value of N-1, it resets to 0 and the output clock clk_out is toggled. The division factor N is specified using a parameter, which can be changed in the instantiation of the module to achieve a different division factor.