Create image_processong.v
This is just an example of image processing on an FPGA, where the input image data is assumed to be 8-bit pixel values and the clock is used to capture the pixel values and perform the image processing operation.

In this example, the grayscale conversion is performed by taking the average of the red, green, and blue components of the input pixel. This code can be extended to perform more complex image processing operations, such as filtering, thresholding, edge detection, etc.