Create FIFO_BUFFER.v
In this example, the FIFO buffer is implemented using an array of depth registers memory to store the data. The write_pointer and read_pointer are used to keep track of the write and read positions in the buffer. The write and read operations are performed in the always block on the positive edge of the clock clk. The reset input initializes the write and read pointers to 0. The write and read operations are performed when the write_enable and read_enable signals are asserted, respectively. The write_full and read_empty outputs indicate the status of the buffer. The read_temp register is used to hold the data to be read until it can be assigned to the read_data output.