Create State_machine.v
This code implements a simple 4-state state machine with inputs clk, reset, start, state_in, and outputs state_out. The next state is determined by a case statement that is triggered by the current state value. The state_out signal is updated on each rising edge of the clock or a reset signal.